+++ trading_core/src/strategies/tpsl_sniper_2/entry/scalp.rs
@@ -535,7 +535,7 @@ mod tests {
             wallet.into(),
             side,
             sol,
-            tokens,
+            tokens as u64,
             format!("sig-{wallet}-{slot}-{secs}"),
             slot,
             base_time() + chrono::Duration::seconds(secs),
@@ -705,7 +705,8 @@ mod tests {
             }
             return Some(EntryFill {
                 price: cand.price_per_token,
-                amount_tokens: cand.token_amount,
+                // EntryFill.amount_tokens is f64; token_amount is exact raw u64.
+                amount_tokens: cand.token_amount as f64,
                 tx_signature: cand.tx_signature.clone(),
                 block_time: cand.block_time,
             });
